Added in Carmen's suggestions.
