+++ client/src/components/SignUp.js
@@ -1,20 +1,9 @@
 import React, { useState, useContext } from "react";
 import { Link } from "react-router-dom";
 import { makeStyles, withStyles } from "@material-ui/core/styles";
-import {
-  Button,
-  TextField,
-  Box,
-  Snackbar,
-  Tooltip,
-  Typography,
-} from "@material-ui/core";
-import MuiAlert from "@material-ui/lab/alert";
+import { Button, TextField, Box, Tooltip, Typography } from "@material-ui/core";
 import UserContext from "../contexts/UserContext";
-
-function Alert(props) {
-  return <MuiAlert elevation={6} variant="filled" {...props} />;
-}
+import WarningSnackbar from "./WarningSnackbar";
 
 const ErrorTooltip = withStyles((theme) => ({
   arrow: {
@@ -88,8 +77,6 @@ function SignUp(props) {
   const [openSnack, setOpenSnack] = useState(false);
   const [snackText, setSnackText] = useState("");
   const value = useContext(UserContext);
-  const vertical = "top";
-  const horizontal = "center";
 
   function sleep(time) {
     return new Promise((resolve) => setTimeout(resolve, time));
@@ -152,24 +139,20 @@ function SignUp(props) {
         .then((response) => response.json())
         .then(function (response) {
           if (response.status === "success") {
-            console.log("Success:", email);
             const loginSuccess = value.handleLogin(email, password);
             if (loginSuccess) props.history.push("/dashboard");
           } else {
             handleSnack(response.message);
-            console.log(response.message);
           }
         })
-        .catch((error) => {
-          console.error("Error:", error);
-        });
+        .catch(() => {});
     }
   };
 
   return (
     <section className={classes.signup}>
       <Box className={classes.formContainer}>
-        <form form onSubmit={handleSignup}>
+        <form onSubmit={handleSignup}>
           <h2 className={classes.h2}>Sign up</h2>
           <label>Your Name</label>
           <TextField
@@ -190,8 +173,6 @@ function SignUp(props) {
             fullWidth
             required
             type="email"
-            // error={!!errors.email}
-            // helperText={errors.email}
             onChange={(e) => setEmail(e.target.value)}
           />
           <label>Password:</label>
@@ -202,8 +183,6 @@ function SignUp(props) {
             fullWidth
             required
             type="password"
-            // error={!!errors.password}
-            // helperText={errors.password}
             inputProps={{ minLength: 6 }}
             onChange={(e) => setPassword(e.target.value)}
           />
@@ -224,13 +203,11 @@ function SignUp(props) {
               required
               type="password"
               error={!!errors.confirm}
-              // helperText={errors.confirm}
               onChange={(e) => setConfirm(e.target.value)}
             />
           </ErrorTooltip>
           <Button
             className={classes.button}
-            // onClick={handleClick}
             type="submit"
             variant="contained"
             color="secondary"
@@ -244,16 +221,11 @@ function SignUp(props) {
             Login
           </Link>
         </Box>
-        <Snackbar
-          open={openSnack}
-          autoHideDuration={6000}
-          onClose={handleCloseSnack}
-          anchorOrigin={{ vertical, horizontal }}
-        >
-          <Alert onClose={handleCloseSnack} severity="error">
-            {snackText}
-          </Alert>
-        </Snackbar>
+        <WarningSnackbar
+          openSnack={openSnack}
+          handleCloseSnack={handleCloseSnack}
+          snackText={snackText}
+        />
       </Box>
     </section>
   );
